extract-files: fallback to DUMP/vendor_boot file for dtb.img

+++ extract-files.sh
@@ -71,7 +71,13 @@ find "${TMP_DIR}/dtbs" -type f -name "*.dtb" \
     -exec basename {} \;
 cp -f "${DUMP}/dtbo.img" "${MY_DIR}/dtbo.img"
 echo "  - dtbo.img"
-cp -f "${DUMP}/vendor_boot/dtb.img" "${MY_DIR}/dtb.img"
+if [ -f "${TMP_DIR}/vendor_boot.out/dtb" ]; then
+    cp -f "${TMP_DIR}/vendor_boot.out/dtb" "${MY_DIR}/dtb.img"
+elif [ -f "${DUMP}/vendor_boot/dtb.img" ]; then
+    cp -f "${DUMP}/vendor_boot/dtb.img" "${MY_DIR}/dtb.img"
+elif [ -f "${DUMP}/vendor_boot" ]; then
+    cp -f "${DUMP}/vendor_boot" "${MY_DIR}/dtb.img"
+fi
 echo "  - dtb.img"
 
 ### Modules
